refactor(tui): extract slash-command parsing to app/commands.rs (S4 PR 1)
S4 phase 1 of the app.rs decomposition. Pure-fn slash-command parsing
chunk lifted into a sibling submodule. No behaviour change; all 84 TUI
tests pass.

Mechanics:
  - rename: crates/memphis-tui/src/app.rs → crates/memphis-tui/src/app/mod.rs
  - new:    crates/memphis-tui/src/app/commands.rs (671 LoC)
  - app/mod.rs: 6039 → 5395 LoC (~640 LoC moved out)

Functions extracted:
  classify_input_route, classify_command_route, is_native_command_tokens,
  extension_host_command_for_tokens, parse_supported_flags,
  legacy_cli_fallback_notice, unsupported_tui_command_notice,
  split_stderr_details, summarize_host_command_error, split_command_tokens

Visibility: AppState callers reach the helpers via `pub(super)` from
app/commands.rs, plus one `pub(crate) use commands::classify_input_route`
re-export to keep main.rs's existing `app::classify_input_route` import
path unchanged.

This is the first of multiple S4 PRs targeting <1500 LoC for app/mod.rs.
Next planned slices: append_*_host_result methods (~46 fns, ~1100 LoC)
into app/host_results.rs, then format/styling helpers.
